+++ faster/src/main/java/ewing/faster/common/DictionaryController.java
@@ -15,7 +15,6 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.math.BigInteger;
 import java.util.List;
 
 /**
@@ -48,7 +47,7 @@ public class DictionaryController {
     @ApiOperation("删除字典项")
     @PostMapping("/deleteDictionary")
     @PreAuthorize("hasAuthority('" + AuthorityCodes.DICTIONARY_DELETE + "')")
-    public ResultMessage<?> deleteDictionary(BigInteger dictionaryId) {
+    public ResultMessage<?> deleteDictionary(Long dictionaryId) {
         dictionaryService.deleteDictionary(dictionaryId);
         return new ResultMessage<>();
     }
