+++ faster/src/main/java/ewing/faster/security/SecurityController.java
@@ -18,7 +18,6 @@ import org.springframework.web.bind.annotation.*;
 
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
-import java.math.BigInteger;
 import java.util.List;
 
 /**
@@ -67,7 +66,7 @@ public class SecurityController {
     @ApiOperation("删除权限")
     @PostMapping("/deleteAuthority")
     @PreAuthorize("hasAuthority('" + AuthorityCodes.AUTHORITY_DELETE + "')")
-    public ResultMessage<?> deleteAuthority(BigInteger authorityId) {
+    public ResultMessage<?> deleteAuthority(Long authorityId) {
         securityService.deleteAuthority(authorityId);
         return new ResultMessage<>();
     }
@@ -141,7 +140,7 @@ public class SecurityController {
     @ApiOperation("删除角色")
     @PostMapping("/deleteRole")
     @PreAuthorize("hasAuthority('" + AuthorityCodes.ROLE_DELETE + "')")
-    public ResultMessage<?> deleteRole(BigInteger roleId) {
+    public ResultMessage<?> deleteRole(Long roleId) {
         securityService.deleteRole(roleId);
         return new ResultMessage<>();
     }
