+++ faster/src/main/java/ewing/faster/security/SecurityUser.java
@@ -5,7 +5,6 @@ import ewing.common.utils.TreeUtils;
 import ewing.faster.security.vo.AuthorityNode;
 import org.springframework.security.core.userdetails.UserDetails;
 
-import java.math.BigInteger;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -17,7 +16,7 @@ import java.util.List;
  */
 public class SecurityUser implements UserDetails {
 
-    private BigInteger userId;
+    private Long userId;
 
     private String username;
 
@@ -57,11 +56,11 @@ public class SecurityUser implements UserDetails {
                 AuthorityNode::setChildren);
     }
 
-    public BigInteger getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(BigInteger userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
